Rename copy question file to copy content (#12136)
* Rename copy question file to copy content

* commit renamed files
